Feature/molecule marker entropy register (#52)
Implemented proper **territory ownership model for organisms**. They
have a molecule marker register (MR) that can be set with new ISA
instructions SMR / SMRI / SMRS. Molecules written into the environment
are marked with this marker. A FORK that the created organism becomes
the owner of all molecules that were marked with the same marker as the
current MR of the FORKing organism.
The molecule bit structure needed to be extended for this model:
| Meaning | Before | Now |
|---|---|---|
| Type | 4bit | 8bit |
| Value | 16bit | 20bit |
| Marker | - | 4bit |

Also a **simple thermodynamic model** was implemented:
Introduced a new entropy register (SR) with a max value configured in
Config.java. If this register exceeds the max value the organism dies.
The register is increased for every energy it uses. To reduce entropy
the organism needs to POKE/PPK new molecules into the environment. The
value of the reduction equals the value of the created molecule.
This is a simple model, but I have a max more flexible in mind, that
will probably also get implemented soon.
For more info look at: #38 

This PR closes: #36 #37